backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-jose[cryptography]==3.3.0
pydantic-settings==2.7.0
python-multipart==0.0.20
stripe==8.11.0
sqlalchemy==2.0.31
resend==2.3.0
